+++ src/generators/nodes/PendingBlock.ts
@@ -1,6 +1,8 @@
 import Node from './shared/Node';
 import Block from '../dom/Block';
+import State from '../dom/State';
 
 export default class PendingBlock extends Node {
 	_block: Block;
+	_state: State;
 }
