+++ src/generators/nodes/ThenBlock.ts
@@ -1,6 +1,8 @@
 import Node from './shared/Node';
 import Block from '../dom/Block';
+import State from '../dom/State';
 
 export default class ThenBlock extends Node {
 	_block: Block;
+	_state: State;
 }
